+++ src/main.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: main.c,v 1.478 2008-08-26 10:48:52 yangtse Exp $
+ * $Id: main.c,v 1.479 2008-08-28 22:41:35 yangtse Exp $
  ***************************************************************************/
 #include "setup.h"
 
@@ -232,7 +232,7 @@ typedef enum {
 #  include <io.h>
 #  include <sys/types.h>
 #  include <sys/stat.h>
-#  define lseek(fdes,offset,whence)  _lseek(fdes, offset, whence)
+#  define lseek(fdes,offset,whence)  _lseek(fdes, (long)offset, whence)
 #  define fstat(fdes,stp)            _fstat(fdes, stp)
 #  define stat(fname,stp)            _stat(fname, stp)
 #  define struct_stat                struct _stat
